move AA drought specific components to AADroughtPanel folder
