+++ frontend/src/components/MapView/LeftPanel/AnticipatoryActionPanel/AnticipatoryActionDroughtPanel/DistrictView/utils.ts
@@ -3,8 +3,8 @@ import {
   AnticipatoryActionState,
 } from 'context/anticipatoryActionStateSlice/types';
 import { getSeason } from 'context/anticipatoryActionStateSlice/utils';
-import { AADataSeverityOrder } from '../utils';
-import { TimelineRow } from '../Timeline/utils';
+import { AADataSeverityOrder } from '../../utils';
+import { TimelineRow } from '../../Timeline/utils';
 
 export function districtViewTransform(
   data: AnticipatoryActionDataRow[] | undefined,
